Add resilient syslog handler
Syslog and application services can be restarted at various times which
means that the syslog server could momentarily be unavailable to an
application at certain times. The normal syslog handler in the python
standard library does not handle this well because of a couple reasons:
 - Syslog server must be available on during application initilization
 - If syslog becomes unavailable during normal runtime, a traceback will
   printed to the console, but the actual log message will not appear
 - Hard to find out which application the logging belong to

By adding and extending on the normal syslog handler from python
standard library, it is possible to handle connection issues to syslog
and fallback to console logging in a controlled manner.